+++ src/main/java/com/didalgo/intellij/chatgpt/chat/models/StandardModel.java
@@ -20,6 +20,7 @@ public enum StandardModel implements ModelType {
 
     GEMINI_1_5_PRO_LATEST("gemini-1.5-pro-latest", ModelFamily.GEMINI, 2097152),
     GEMINI_1_5_FLASH_LATEST("gemini-1.5-flash-latest", ModelFamily.GEMINI, 1048576),
+    GEMINI_2_0_FLASH("gemini-2.0-flash", ModelFamily.GEMINI, 1048576),
 
     GPT_3_5_TURBO("gpt-3.5-turbo", ModelFamily.OPEN_AI, 16385),
     GPT_3_5_TURBO_0301("gpt-3.5-turbo-0301", ModelFamily.OPEN_AI, 4096),
@@ -41,6 +42,10 @@ public enum StandardModel implements ModelType {
     GPT_4_1106_PREVIEW("gpt-4-1106-preview", ModelFamily.OPEN_AI, 128000),
     GPT_4_0125_PREVIEW("gpt-4-0125-preview", ModelFamily.OPEN_AI, 128000),
 
+    O3_MINI("o3-mini", ModelFamily.OPEN_AI, 200000, false, false, IncompatibleChatOptions.O1_OVERRIDE),
+    O3_MINI_2025_01_31("o3-mini-2025-01-31", ModelFamily.OPEN_AI, 200000, false, false, IncompatibleChatOptions.O1_OVERRIDE),
+    O1("o1", ModelFamily.OPEN_AI, 200000, false, false, IncompatibleChatOptions.O1_OVERRIDE),
+    O1_2024_12_17("o1-2024-12-17", ModelFamily.OPEN_AI, 200000, false, false, IncompatibleChatOptions.O1_OVERRIDE),
     O1_MINI("o1-mini", ModelFamily.OPEN_AI, 128000, false, false, IncompatibleChatOptions.O1_OVERRIDE),
     O1_MINI_2024_09_12("o1-mini-2024-09-12", ModelFamily.OPEN_AI, 128000, false, false, IncompatibleChatOptions.O1_OVERRIDE),
     O1_PREVIEW("o1-preview", ModelFamily.OPEN_AI, 128000, false, false, IncompatibleChatOptions.O1_OVERRIDE),